feat: Add API client and update index.ts
